+++ diskann-disk/Cargo.toml
@@ -33,7 +33,6 @@ cfg-if.workspace = true
 futures-util = { workspace = true, default-features = false }
 hashbrown.workspace = true
 half = { workspace = true, features = ["bytemuck", "num-traits"] }
-diskann-platform = { workspace = true }
 rand.workspace = true
 rayon.workspace = true
 serde = { workspace = true, features = ["derive"] }
@@ -46,8 +45,24 @@ vfs = { workspace = true }
 opentelemetry = { workspace = true, optional = true }
 
 [target.'cfg(target_os = "linux")'.dependencies]
-io-uring = "0.6.4"
+io-uring = "0.7.12"
+
+[target.'cfg(unix)'.dependencies]
 libc = "0.2.148"
+
+[target.'cfg(target_os = "windows")'.dependencies.windows-sys]
+version = "0.61.2"
+features = [
+    "Win32_Foundation",
+    "Win32_Security",
+    "Win32_Storage",
+    "Win32_Storage_FileSystem",
+    "Win32_System_IO",
+    "Win32_System_Threading",
+    "Win32_System_WindowsProgramming",
+    "Win32_System_ProcessStatus",
+    "Win32_System_SystemInformation"
+]
 
 [dev-dependencies]
 rstest.workspace = true
