+++ diskann-disk/benches/benchmarks/aligned_file_reader_bench.rs
@@ -5,7 +5,7 @@
 use std::time::Duration;
 
 use criterion::Criterion;
-use diskann_disk::utils::aligned_file_reader::{
+use diskann_disk::search::provider::aligned_file_reader::{
     traits::{AlignedFileReader, AlignedReaderFactory},
     AlignedFileReaderFactory, AlignedRead,
 };
