+++ diskann-disk/src/search/provider/aligned_file_reader/platform/windows/ssd_io_context.rs
@@ -3,18 +3,17 @@
  * Licensed under the MIT license.
  */
 
-use std::fs::File;
-
-use io_uring::IoUring;
+use super::{FileHandle, IOCompletionPort};
 
 // The IOContext struct for disk I/O. One for each thread.
+#[derive(Default)]
 pub struct IOContext {
-    pub file_handle: File,
-    pub ring: IoUring,
+    pub file_handle: FileHandle,
+    pub io_completion_port: IOCompletionPort,
 }
 
 impl IOContext {
-    pub fn new(file_handle: File, ring: IoUring) -> Self {
-        IOContext { file_handle, ring }
+    pub fn new() -> Self {
+        Self::default()
     }
 }
