+++ diskann-disk/src/search/provider/aligned_file_reader/traits/mod.rs
@@ -3,9 +3,8 @@
  * Licensed under the MIT license.
  */
 
-#[allow(clippy::module_inception)]
 mod aligned_file_reader;
 pub use aligned_file_reader::AlignedFileReader;
 
-pub mod aligned_reader_factory;
+mod aligned_reader_factory;
 pub use aligned_reader_factory::AlignedReaderFactory;
