+++ diskann-disk/src/utils/instrumentation/mod.rs
@@ -4,3 +4,6 @@
  */
 mod perf_logger;
 pub use perf_logger::{BuildMergedVamanaIndexCheckpoint, DiskIndexBuildCheckpoint, PerfLogger};
+
+mod timer;
+pub use timer::Timer;
