+++ diskann-disk/src/utils/instrumentation/timer/macos.rs
