--- diskann-platform/Cargo.toml
@@ -1,44 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-[package]
-name = "diskann-platform"
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-repository.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tracing.workspace = true
-
-[target.'cfg(target_os = "linux")'.dependencies]
-io-uring = "0.6.4"
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2.148"
-
-[target.'cfg(target_os = "windows")'.dependencies.windows-sys]
-version = "0.59"
-features = [
-    "Win32_Foundation",
-    "Win32_Security",
-    "Win32_Storage",
-    "Win32_Storage_FileSystem",
-    "Win32_System_IO",
-    "Win32_System_Threading",
-    "Win32_System_WindowsProgramming",
-    "Win32_System_ProcessStatus",
-    "Win32_System_SystemInformation"
-]
-
-[package.metadata.docs.rs]
-default-target = "x86_64-pc-windows-msvc"
-targets = ["x86_64-unknown-linux-gnu", "aarch64-pc-windows-msvc", "x86_64-pc-windows-msvc"]
-
-[lints]
-# Commenting out for now. It would give warnings.
-# clippy.undocumented_unsafe_blocks = "warn"
-clippy.uninlined_format_args = "allow"
