--- diskann-platform/src/lib.rs
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT license.
- */
-
-#[cfg(target_os = "windows")]
-pub mod win;
-
-#[cfg(target_os = "windows")]
-pub use win::*;
-
-#[cfg(target_os = "linux")]
-pub mod linux;
-
-#[cfg(target_os = "linux")]
-pub use linux::*;
-
-#[cfg(target_os = "macos")]
-pub mod macos;
-
-#[cfg(target_os = "macos")]
-pub use macos::*;
