+++ diskann-tools/Cargo.toml
@@ -46,7 +46,7 @@ diskann-utils = { workspace = true, features = ["testing"] }
 [features]
 default = []
 # When a feature is enabled in tools, enable it in dependencies
-perf_test = ["diskann-providers/perf_test"]
+perf_test = ["diskann-disk/perf_test"]
 
 [lints]
 clippy.undocumented_unsafe_blocks = "warn"
